change mistaken names of folders and files
